Convert ToolbarDivider into internal component
