+++ docs/data/data-grid/components/toolbar/toolbar.md
@@ -1,6 +1,6 @@
 ---
 productId: x-data-grid
-components: Toolbar, ToolbarButton, ToolbarDivider
+components: Toolbar, ToolbarButton
 packageName: '@mui/x-data-grid'
 githubLabel: 'scope: data grid'
 ---
@@ -26,8 +26,6 @@ import { Toolbar, ToolbarButton, ToolbarDivider } from '@mui/x-data-grid';
 
 <Toolbar>
   <ToolbarButton />
-  <ToolbarDivider />
-  <ToolbarButton />
 </Toolbar>;
 ```
 
@@ -40,11 +38,6 @@ It renders a styled `<div />` element.
 
 `<ToolbarButton />` is a button for performing actions from the toolbar.
 It renders the `baseIconButton` slot.
-
-### Toolbar Divider
-
-`<ToolbarDivider />` is a component that separates toolbar items.
-It renders the `baseDivider` slot.
 
 ## Recipes
