+++ docs/pages/x/api/charts/bar-chart.json
@@ -133,7 +133,6 @@
       "class": null
     },
     { "name": "baseButton", "description": "", "class": null },
-    { "name": "baseDivider", "description": "", "class": null },
     { "name": "baseIconButton", "description": "", "class": null },
     {
       "name": "legend",
