+++ docs/pages/x/api/charts/funnel-chart.json
@@ -105,7 +105,6 @@
       "class": null
     },
     { "name": "baseButton", "description": "", "class": null },
-    { "name": "baseDivider", "description": "", "class": null },
     { "name": "baseIconButton", "description": "", "class": null },
     {
       "name": "funnelSection",
