+++ docs/pages/x/api/charts/line-chart.json
@@ -119,7 +119,6 @@
       "class": null
     },
     { "name": "baseButton", "description": "", "class": null },
-    { "name": "baseDivider", "description": "", "class": null },
     { "name": "baseIconButton", "description": "", "class": null },
     {
       "name": "legend",
