+++ docs/pages/x/api/charts/pie-chart.json
@@ -52,7 +52,6 @@
   ],
   "slots": [
     { "name": "baseButton", "description": "", "class": null },
-    { "name": "baseDivider", "description": "", "class": null },
     { "name": "baseIconButton", "description": "", "class": null },
     {
       "name": "legend",
