+++ docs/pages/x/api/charts/scatter-chart.json
@@ -124,7 +124,6 @@
       "class": null
     },
     { "name": "baseButton", "description": "", "class": null },
-    { "name": "baseDivider", "description": "", "class": null },
     { "name": "baseIconButton", "description": "", "class": null },
     {
       "name": "legend",
