+++ docs/translations/api-docs/charts/funnel-chart/funnel-chart.json
@@ -63,7 +63,6 @@
     "axisTick": "Custom component for the axis tick.",
     "axisTickLabel": "Custom component for tick label.",
     "baseButton": "",
-    "baseDivider": "",
     "baseIconButton": "",
     "funnelSection": "Custom component for funnel section.",
     "funnelSectionLabel": "Custom component for funnel section label.",
