+++ docs/translations/api-docs/charts/pie-chart/pie-chart.json
@@ -41,7 +41,6 @@
   "classDescriptions": {},
   "slotDescriptions": {
     "baseButton": "",
-    "baseDivider": "",
     "baseIconButton": "",
     "legend": "Custom rendering of the legend.",
     "loadingOverlay": "Overlay component rendered when the chart is in a loading state.",
