+++ docs/translations/api-docs/charts/scatter-chart/scatter-chart.json
@@ -79,7 +79,6 @@
     "axisTick": "Custom component for the axis tick.",
     "axisTickLabel": "Custom component for tick label.",
     "baseButton": "",
-    "baseDivider": "",
     "baseIconButton": "",
     "legend": "Custom rendering of the legend.",
     "loadingOverlay": "Overlay component rendered when the chart is in a loading state.",
