+++ packages/x-charts/src/Toolbar/index.ts
@@ -1,5 +1,4 @@
 export * from './Toolbar';
 export * from './ToolbarButton';
-export * from './ToolbarDivider';
 export * from './Toolbar.types';
 export * from './chartToolbarClasses';
