+++ packages/x-charts/src/models/slots/chartsBaseSlots.ts
@@ -1,12 +1,7 @@
 import * as React from 'react';
-import {
-  ChartBaseButtonProps,
-  ChartBaseDividerProps,
-  ChartBaseIconButtonProps,
-} from './chartsBaseSlotProps';
+import { ChartBaseButtonProps, ChartBaseIconButtonProps } from './chartsBaseSlotProps';
 
 export interface ChartsBaseSlots {
   baseButton: React.ComponentType<ChartBaseButtonProps>;
   baseIconButton: React.ComponentType<ChartBaseIconButtonProps>;
-  baseDivider: React.ComponentType<ChartBaseDividerProps>;
 }
